+++ src/lib/health-checks/index.ts
@@ -9,15 +9,18 @@ export {
 
 export {
   checkAnthropicHealth,
-  checkPosthogOverallHealth,
-  checkPosthogComponentHealth,
   checkGithubHealth,
   checkNpmOverallHealth,
   checkNpmComponentHealth,
   checkCloudflareOverallHealth,
   checkCloudflareComponentHealth,
-  resetPosthogHealthCache,
 } from './statuspage';
+
+export {
+  checkPosthogOverallHealth,
+  checkPosthogComponentHealth,
+  resetPosthogHealthCache,
+} from './incidentio';
 
 export {
   checkLlmGatewayHealth,
